feat: Initial code
This initial script provides a few extra capabilities on top of the usual "blinky". It sets up USB CDC ACM for logging, which will cause the device to show up as a serial port on most machines. It then hooks up the logging subsystem to use the USB interface. Finally, it also uses binary descriptors to grab the build timestamp.